Fix: 테스트 도메인 변경
